drivers: console: uart_console: handle idle PM
Allow the uart_console console_out implementation to integrate with
'Device Idle Power Management'.

Without using idle PM, it is impossible for the application to manage
the power state of a UART instance, given that any module in the
application could theoretically LOG_* or printk at any point in time.

This removes the need to manually call device_set_power_state when the
lowest power states must be reached, while still allowing console and
printk messages to be output.

Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>